feat: group delete 할 때 inviteToken을 받아서 일치하는지 검

+++ src/main/java/com/insighton/core/domain/groups/service/impl/GroupServiceImpl.java
@@ -6,6 +6,7 @@ import com.insighton.core.domain.groups.dto.response.GroupAdminResponse;
 import com.insighton.core.domain.groups.dto.response.GroupResponse;
 import com.insighton.core.domain.groups.entity.Group;
 import com.insighton.core.domain.groups.exception.GroupNotFoundException;
+import com.insighton.core.domain.groups.exception.InvitationTokenMismatchException;
 import com.insighton.core.domain.groups.exception.InviteTokenNotFoundException;
 import com.insighton.core.domain.groups.exception.UnAuthorizedAccessException;
 import com.insighton.core.domain.groups.repository.GroupRepository;
@@ -16,6 +17,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
@@ -92,9 +94,13 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     @Transactional
-    public void deleteGroup(Long groupId) {
+    public void deleteGroup(Long groupId, String inviteToken) {
         // 대상 그룹 조회 (없을 시 exception 던지기)
         Group groupEntity = groupFindById(groupId);
+
+        if (!Objects.equals(groupEntity.getInviteToken(), inviteToken)) {
+            throw new InvitationTokenMismatchException(inviteToken);
+        }
 
         groupRepository.delete(groupEntity);
     }
